requirements.txt
azure-functions
azure-storage-blob
pillow
image
azure-cosmosdb-table
flask
python-dotenv
python-jose
flask-cors
six